% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_result.R
\name{stand_result}
\alias{stand_result}
\title{convert different statistical reports into a standardized framework (Cohen's D and Variance/Standard Error of Cohen's D)}
\usage{
stand_result(eff_type, raw_effect_size, sd, n_t, n_c)
}
\arguments{
\item{eff_type}{The effect type to convert. Currently accepted options are
"d_i_d" (difference in differences), "d_i_m" (difference in means),
'd' (For Cohen's D -- this won't recalculate Cohen's D but will estimate
its variance and standard error), "unspecified null" (for when authors report
that they found null effects but don't give a more precise recording of what
those were), "eta_squared" (for eta squared), "reg_coef" (for a regression
coefficient), "t_test" (for a T statistic), and "f_test" (for an F test).}

\item{raw_effect_size}{the reported test statistic, in terms of the reported
statistical information. So, if the eff_type = "t_test", raw_effect_size will
equal the T statistic.}

\item{sd}{the standard deviation by which you are standardizing the given test statistic.
This is needed for some but not all of the equations -- F test and T '
statistics already contain information about samplng distribution and thus
can be converted directly and this parameter can be left blank.}

\item{n_t}{sample size in the treatment group.}

\item{n_c}{sample size in the control group.}
}
\description{
convert different statistical reports into a standardized framework (Cohen's D and Variance/Standard Error of Cohen's D)
}
\examples{
stand_result(eff_type = 'd_i_m', raw_effect_size = 2.43,  sd = 6.3, n_t = 100, n_c = 100)
}
